Exclude tests from build output to fix build. Also move types into src to avoid generation of scenario.js on build which could end up out of sync
