--- types/models/scenario.js
@@ -1,2 +1,0 @@
-'use strict'
-Object.defineProperty(exports, '__esModule', { value: true })
